requirements.txt
azure-core==1.25.0
azure-storage-blob==12.13.1
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_884c889c-96af-444f-bd6d-daddb5e9a462ykj3l5n_/croots/recipe/certifi_1655968814730/work/certifi
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
idna==3.3
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.0
pycparser==2.21
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.11
